# models/cus_array_string.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/swag"
)

// CusArrayString cus array string
// swagger:model CusArrayString
type CusArrayString struct {

	// set
	Set bool `json:"Set,omitempty"`

	// value
	Value []string `json:"Value"`
}

// Validate validates this cus array string
func (m *CusArrayString) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CusArrayString) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CusArrayString) UnmarshalBinary(b []byte) error {
	var res CusArrayString
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}