# models/cus_int64.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/swag"
)

// CusInt64 cus int64
// swagger:model CusInt64
type CusInt64 struct {

	// set
	Set bool `json:"Set,omitempty"`

	// value
	Value int64 `json:"Value,omitempty"`
}

// Validate validates this cus int64
func (m *CusInt64) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CusInt64) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CusInt64) UnmarshalBinary(b []byte) error {
	var res CusInt64
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}