# models/cus_string.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/swag"
)

// CusString cus string
// swagger:model CusString
type CusString struct {

	// set
	Set bool `json:"Set,omitempty"`

	// value
	Value string `json:"Value,omitempty"`
}

// Validate validates this cus string
func (m *CusString) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CusString) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CusString) UnmarshalBinary(b []byte) error {
	var res CusString
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}